Merged PR 454: Bug Fixes for BI Application - Pre-Submission
Bug 9037: Fixed task status displaying 'COMPLETE' instead of 'COMPLETED' in the BI Application - Summary of Personal Details page.
Bug 9038: Corrected the status for the Alternative Email Address task in the BI Application - Summary of Personal Details page.
Bug 8851: Resolved the issue with the incorrect display name for a button in the Capture Mobile Phone to Send OTPC section of the BI Application - Pre-submission page.
Bug 8864: Fixed the wrong error message that appeared when entering an invalid telephone number in the Capture Mobile Phone to Send OTPC section of the BI Application - Pre-submission page.

Related work items: #8851, #8864, #9037, #9038